oxo-loader/requirements.txt
certifi==2018.1.18
chardet==3.0.4
configparser==3.5.0
idna==2.6
pymysql==0.9.3
neo4j==4.4.3
PyYAML==5.3.1
requests==2.27.1
urllib3==1.26.8
pandas==1.4.1
